# docs/api/tcheck-corporate.openapi.yaml
openapi: 3.0.3
info:
  title: TallyCheck Corporate API
  description: |
    Corporate attendance, SafeChild pickup, BLE beacons and tenant settings.

    This is one of several TallyCheck service contracts. It does **not** cover the
    academic domain (courses, cohorts, enrolment) — that lives in `Tcheck-Backend`
    and is described by `tcheck-academic.openapi.yaml`. Visitor management lives in
    `vcheck.openapi.yaml`. The three are separate documents on purpose: they run on
    different hosts, use different auth, and have colliding paths (`/attendance`
    means corporate clock-in here and class attendance in the academic service).

    ## Multi-tenancy
    Every request resolves to exactly one tenant schema before any handler runs.
    See `shared/tenancy.yaml`. Callers holding a token send nothing extra — the
    token carries the tenant.

    ## Roles
    Roles are enumerated once, in `shared/primitives.yaml#/components/schemas/RoleKey`,
    and must match `libs/auth/src/roles.ts` and `backend/helpers/auth_helper.py`.
    Some tenants still carry the legacy names `hr` and `manager`; the backend
    guards accept both.

    ## Authentication
    All endpoints except `GET /health` and
    `GET /auth/organization-by-subdomain/{subdomain}` require a Bearer JWT.

    ## Response shape
    Successful responses are enveloped as `{ "message": ..., "data": ... }`, with
    `meta` added on paginated collections. See `shared/envelope.yaml`.
  version: 2.1.0
servers:
...
security:
  - Auth0: []

# ==========================================================
# REUSABLE STRUCTURES
# ==========================================================
components:
  securitySchemes:
    Auth0:
      type: openIdConnect
      # Must match the tenant configured in apps/intranet/src/app/app.config.ts.
      # This previously named a different tenant than the app actually used.
      openIdConnectUrl: https://dev-x506paw8afbw6jgu.us.auth0.com/.well-known/openid-configuration

  # ==========================================================
  # REUSABLE SCHEMAS
  # ==========================================================
  schemas:
...
    ErrorResponse:
      $ref: './shared/errors.yaml#/components/schemas/ErrorResponse'
...
    Employee:
      type: object
      properties:
        id:
          type: integer
          example: 12
        email:
          type: string
          example: "philomena.shirandula@adept-techno.com"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Doe"
        role_id:
          type: integer
          example: 1
        role_name:
          $ref: './shared/primitives.yaml#/components/schemas/RoleName'
        department_id:
          type: integer
          nullable: true
          example: 3
        department_name:
          type: string
          nullable: true
          example: "software_development"
        is_active:
          type: boolean
          example: true
        is_approved:
          type: boolean
          example: true
        is_manager:
          type: boolean
          example: false
        hire_date:
          type: string
          format: date
          nullable: true
          example: "2025-03-01"
        created_at:
          type: string
          format: date-time
        standard_shift:
          type: string
          enum: [morning, afternoon]
          example: "morning"
        shift_type:
          type: string
          enum: [standard, extended]
          example: "extended"
        shift_hours:
          type: string
          enum: ["7am-5pm", "9am-7pm", "7am-430pm", "9am-630pm", "custom"]
          example: "9am-7pm"
        shift_duration_hours:
          type: number
          format: float
          example: 9.5
        custom_shift_start:
          type: string
          nullable: true
          example: "08:00"
        custom_shift_end:
          type: string
          nullable: true
          example: "17:30"
...
    Department:
      type: object
      description: >
        A node in the tenant org tree. Despite the name this covers every level —
        regions, sites, divisions, departments, teams — discriminated by
        `unit_type`.
      properties:
        id:
          type: integer
          example: 3
        name:
          type: string
          description: >
            Display text. `display_name` when one was set, otherwise the slug
            prettified. Prefer this for rendering.
          example: "Software Development"
        slug:
          type: string
          description: >
            Stored identifier, and the uniqueness key. Unique among siblings, not
            tenant-wide, so two sites may each have a "reception".
          example: "software_development"
        display_name:
          type: string
          nullable: true
          description: >
            Human-authored label. Exists because the slug round-trip is lossy —
            "iOS Team" stores as "ios_team" and reads back as "Ios Team".
          example: "Software Development"
        unit_type:
          $ref: "#/components/schemas/UnitType"
        manager_id:
          type: integer
          nullable: true
          example: 7
        manager_name:
          type: string
          nullable: true
          example: "Barack Osama"
        employee_count:
          type: integer
          example: 12
        child_count:
          type: integer
          description: Immediate children only, not the whole subtree.
          example: 4
        parent_department_id:
          type: integer
          nullable: true
          example: 1
        parent_department_name:
          type: string
          nullable: true
          example: "Executive"
    DepartmentRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Slugified server-side (lowercased, spaces to underscores).
          example: "finance"
        display_name:
          type: string
          nullable: true
          description: Optional human-authored label, stored verbatim.
          example: "Finance"
        unit_type:
          $ref: "#/components/schemas/UnitType"
        manager_id:
          type: integer
          nullable: true
          example: 7
        parent_department_id:
          type: integer
          nullable: true
          description: >
            Rejected if it does not exist, sits inside this unit's own subtree
            (which would create a loop), or would nest the tree beyond 6 levels.
          example: 2

    # ==========================================================
    # ORGANIZATIONS (MULTI-TENANCY)
    # ==========================================================
...
    ReportEmployeeAttendance:
      type: object
      properties:
        employee_id:
          type: integer
          example: 12
        employee_name:
          type: string
          example: "Nyambane Mogaka"
        department:
          type: string
          example: "software_development"
        total_days_present:
          type: integer
          example: 18
        total_days_absent:
          type: integer
          example: 2
        total_hours_worked:
          type: number
          format: float
          example: 162.5
        avg_clock_in:
          type: string
          example: "08:07"
        avg_clock_out:
          type: string
          example: "17:23"
        late_days:
          type: integer
          example: 3
    ReportDepartmentAttendance:
      type: object
      properties:
        department_id:
          type: integer
          example: 2
        department_name:
          type: string
          example: "Software Development"
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        total_employees:
          type: integer
          example: 4
        avg_attendance_rate:
          type: number
          format: float
          description: "Percentage 0–100"
          example: 94.5
        total_hours_worked:
          type: number
          format: float
          example: 640.5
        employees:
          type: array
          items:
            $ref: "#/components/schemas/ReportEmployeeAttendance"
...
    UnitType:
      type: string
      description: |
        Structural level in the org tree — deliberately NOT a vertical noun.

        `site` covers campus, branch, depot, facility and store; the
        customer-facing word is a display concern the frontend maps per
        organization type. Adding `campus`/`depot`/`ward` here would make every
        new customer a schema change, and the enum would grow without end.
      enum:
        - root
        - region
        - site
        - division
        - department
        - team
      default: department
    DepartmentTreeNode:
      allOf:
        - $ref: "#/components/schemas/Department"
        - type: object
          properties:
            children:
              type: array
              description: Nested child units. Empty for a leaf.
              items:
                $ref: "#/components/schemas/DepartmentTreeNode"


    # ==========================================================
    # ZONES (indoor location)
    # ==========================================================
...
  responses:
    Unauthorized: # Can be referenced as '#/components/responses/unauthorized'
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Business rule conflict (e.g. already clocked in)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Validation error e.g. missing required field
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ==========================================================
  # REUSABLE PARAMETERS
  # ==========================================================
  parameters:
...
    DateFromParam:
      name: date_from
      in: query
      schema:
        type: string
        format: date
      example: "2026-06-01"
    DateToParam:
      name: date_to
      in: query
      schema:
        type: string
        format: date
      example: "2026-06-30"
...
paths:
...
  /departments:
    get:
      tags: [HR - Departments]
      summary: List all departments
      operationId: list_departments
      responses:
        "200":
          description: All departments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Department"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [HR - Departments]
      summary: Create a department (HR only)
      operationId: create_department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentRequest"
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /departments/{id}:
    put:
      tags: [HR - Departments]
      summary: Update a department (HR only)
      operationId: update_department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentRequest"
      responses:
        "200":
          description: Department updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [HR - Departments]
      summary: Delete a department (HR only)
      operationId: delete_department
      description: |
        A department can only be deleted if it has no active employees.
        Attempting to delete a populated department returns `409`.

        If the unit has child units the request is rejected unless
        `reparent_to` is supplied — silently promoting a campus's children to
        roots would lose the org chart. Children are moved in the same
        transaction as the delete.

        **Roles:** HR, Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: reparent_to
          in: query
          required: false
          description: >
            Where to move child units. A department id, or `root` to promote them
            to top level. Required when the unit has children.
          schema:
            type: string
            example: "4"
      responses:
        "200":
          description: Department deleted
          content:
            application/json:
              schema:
                $ref: './shared/envelope.yaml#/components/schemas/DeletedEnvelope'
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Department has active employees
          $ref: "#/components/responses/Conflict"

  # ==========================================================
  # REPORTS
  # ==========================================================
...
  /reports/attendance/departments:
    get:
      tags: [Reports]
      summary: Get attendance report for all departments (HR only)
      operationId: report_attendance_all_departments
      parameters:
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
      responses:
        "200":
          description: Departments attendance records found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportDepartmentAttendance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /reports/attendance/departments/{id}:
    get:
      tags: [Reports]
      summary: Get attendance report for one department (HR only)
      operationId: report_attendance_one_department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - $ref: "#/components/parameters/DateFromParam"
        - $ref: "#/components/parameters/DateToParam"
      responses:
        "200":
          description: Department attendance record found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ReportDepartmentAttendance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /departments/tree:
    get:
      tags: [HR - Departments]
      summary: The org chart, nested
      description: >
        Every unit in the tenant, assembled into `children` lists with roots
        first. A unit whose parent is absent from the result is returned as a
        root rather than dropped, so a partial or filtered view still renders.
      responses:
        "200":
          description: Nested org tree.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DepartmentTreeNode"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /departments/{id}/employees:
    get:
      tags: [HR - Departments]
      summary: Employees in a unit, optionally including everything beneath it
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: include_descendants
          in: query
          required: false
          description: >
            When true, returns everyone in the whole subtree — "show me everyone
            at Nairobi Campus" spans its divisions and teams.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Employees in scope.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './shared/envelope.yaml#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Employee"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==========================================================
  # ZONES
  # ==========================================================
...